Introduce FluentNavBase
