+++ src/Core/Components/Nav/FluentNavCategory.razor
@@ -1,6 +1,6 @@
 ﻿@namespace Microsoft.FluentUI.AspNetCore.Components
 @using Microsoft.FluentUI.AspNetCore.Components.Extensions
-@inherits FluentComponentBase
+@inherits FluentNavBase
 
 <CascadingValue Name="Category" Value="this">
 	<button id="@Id"
