+++ src/Core/Components/Nav/FluentNavItem.razor
@@ -1,6 +1,6 @@
 ﻿@namespace Microsoft.FluentUI.AspNetCore.Components
 @using Microsoft.FluentUI.AspNetCore.Components.Extensions
-@inherits FluentComponentBase
+@inherits FluentNavBase
 
 @if (!string.IsNullOrEmpty(Href))
 {
